feat(small): Repair PR #6611: Add missing dependencies and tests (#7144)
