+++ package.json
@@ -80,6 +80,7 @@
     "js-cookie": "^3.0.5",
     "lodash.isequal": "^4.5.0",
     "lodash.throttle": "^4.1.1",
+    "lucide-react": "0.473.0",
     "next": "16.1.6",
     "next-auth": "^4.24.13",
     "notistack": "^3.0.2",
@@ -181,7 +182,8 @@
       "cookie@<0.7.0": ">=0.7.0",
       "on-headers@<1.1.0": ">=1.1.0",
       "send@<0.19.0": ">=0.19.0",
-      "next@>=16.0.0-canary.0 <16.0.7": ">=16.0.7"
+      "next@>=16.0.0-canary.0 <16.0.7": ">=16.0.7",
+      "zod": "^4.3.6"
     }
   }
 }
